feat: manage brands permission

+++ app/Http/Controllers/BrandController.php
@@ -5,13 +5,12 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreBrandRequest;
 use App\Http\Requests\UpdateBrandRequest;
 use App\Models\Brand;
-use App\Models\Category;
 
 class BrandController extends Controller
 {
     public function index()
     {
-        $this->authorize('viewAny', Category::class);
+        $this->authorize('viewAny', Brand::class);
         $brands = Brand::latest()->get();
 
         return view('brand.index', compact('brands'));
@@ -19,14 +18,14 @@ class BrandController extends Controller
 
     public function create()
     {
-        $this->authorize('create', Category::class);
+        $this->authorize('create', Brand::class);
 
         return view('brand.create');
     }
 
     public function store(StoreBrandRequest $request)
     {
-        $this->authorize('create', Category::class);
+        $this->authorize('create', Brand::class);
         $filePath = optional($request->file('image'))->store('images', ['disk' => 'public_storage']);
 
         Brand::create([
@@ -42,14 +41,15 @@ class BrandController extends Controller
 
     public function edit(Brand $brand)
     {
-//        $this->authorize('update', $brand);
+        $this->authorize('update', $brand);
 
         return view('brand.edit', compact('brand'));
     }
 
     public function update(UpdateBrandRequest $request, Brand $brand)
     {
-//        $this->authorize('update', $brand);
+        $this->authorize('update', $brand);
+
         $filePath = $brand->image_url;
         if ($request->file('image')) {
             $filePath = optional($request->file('image'))->store('images', ['disk' => 'public_storage']);
@@ -68,7 +68,7 @@ class BrandController extends Controller
 
     public function destroy(Brand $brand)
     {
-//        $this->authorize('delete', $brand);
+        $this->authorize('delete', $brand);
 
         return $brand->delete();
     }
